Delete deps that were only being used transitively
The superstore, superstore-sync, and ftdomdelegate packages were being
installed as dependencies for n-ui-foundations, but are not actually
used anywhere in the code. Rather than this being a simple case of
deleting superfluous dependencies, however, this is a holdover feature
from the project's time as a bower dependency, where those three
dependencies could be installed and used via this package. Using
transitive dependencies like this is inappropriate and unstable: it
makes the dependency chain hard to follow/audit, and leaves a dependee
subject to breaking if any of the transitive dependencies are updated or
removed from n-ui-foundations. Instead, they should be installed as
direct dependencies of packages that need them.